+++ src/elements/Button/Button.css
@@ -31,6 +31,7 @@
 
 .secondary {
   composes: base;
+  background-image: linear-gradient(to bottom, rgba(0,0,0,.0), rgba(0,0,0,.05) 100%);
   border: 1px solid var(--border);
 }
